Componentize dirty code modal
